+++ frontend/components/actions/DeviceCalibrationActionForm.vue
@@ -1,8 +1,10 @@
 <!--
-SPDX-FileCopyrightText: 2020 - 2024
+SPDX-FileCopyrightText: 2020 - 2026
 - Nils Brinckmann <nils.brinckmann@gfz-potsdam.de>
 - Marc Hanisch <marc.hanisch@gfz-potsdam.de>
+- Rubankumar Moorthy <r.moorthy@fz-juelich.de>
 - Helmholtz Centre Potsdam - GFZ German Research Centre for Geosciences (GFZ, https://www.gfz-potsdam.de)
+- Research Centre Juelich GmbH - Institute of Bio- and Geosciences Agrosphere (IBG-3, https://www.fz-juelich.de/en/ibg/ibg-3)
 
 SPDX-License-Identifier: EUPL-1.2
 -->
@@ -211,6 +213,10 @@ export default class DeviceCalibationActionForm extends Vue {
     this.createActionCopy(this.value)
   }
 
+  mounted () {
+    this.defaultCurrentCalibrationDate()
+  }
+
   setCurrentCalibrationDateAndValidate (aDate: DateTime | null) {
     this.actionCopy.currentCalibrationDate = aDate
     if (this.actionCopy.nextCalibrationDate !== null) {
@@ -337,9 +343,24 @@ export default class DeviceCalibationActionForm extends Vue {
     this.actionCopy = DeviceCalibrationAction.createFromObject(action)
   }
 
-  @Watch('value', { immediate: true, deep: true })
+  @Watch('value')
   onValueChanged (val: DeviceCalibrationAction) {
     this.createActionCopy(val)
+  }
+
+  /**
+   * Auto-fill Current calibration date (UTC, minute).
+   * We don't run validate() here; we just clear any stale error UI.
+   */
+  private defaultCurrentCalibrationDate () {
+    if (!this.actionCopy.currentCalibrationDate) {
+      this.actionCopy.currentCalibrationDate = DateTime.utc().startOf('minute')
+      this.$emit('input', this.actionCopy)
+      this.$nextTick(() => {
+        const formRef = this.$refs.datesForm as (Vue & { resetValidation?: () => void }) | undefined
+        formRef?.resetValidation?.()
+      })
+    }
   }
 }
 </script>