+++ frontend/components/configurations/dynamicLocation/DynamicLocationWizard.vue
@@ -1,10 +1,12 @@
 <!--
-SPDX-FileCopyrightText: 2020 - 2023
+SPDX-FileCopyrightText: 2020 - 2026
 - Nils Brinckmann <nils.brinckmann@gfz-potsdam.de>
 - Marc Hanisch <marc.hanisch@gfz-potsdam.de>
 - Tobias Kuhnert <tobias.kuhnert@ufz.de>
 - Tim Eder <tim.eder@ufz.de>
+- Rubankumar Moorthy <r.moorthy@fz-juelich.de>
 - Helmholtz Centre Potsdam - GFZ German Research Centre for Geosciences (GFZ, https://www.gfz-potsdam.de)
+- Research Centre Juelich GmbH - Institute of Bio- and Geosciences Agrosphere (IBG-3, https://www.fz-juelich.de/en/ibg/ibg-3)
 
 SPDX-License-Identifier: EUPL-1.2
 -->
@@ -107,7 +109,7 @@ SPDX-License-Identifier: EUPL-1.2
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from 'nuxt-property-decorator'
+import { Component, Prop, Vue, Watch } from 'nuxt-property-decorator'
 import DynamicLocationActionDateForm
   from '@/components/configurations/dynamicLocation/DynamicLocationActionDateForm.vue'
 import { DynamicLocationAction } from '@/models/DynamicLocationAction'
@@ -164,6 +166,20 @@ export default class DynamicLocationWizard extends Vue {
 
   update (updatedLocationAction: DynamicLocationAction) {
     this.$emit('input', updatedLocationAction)
+  }
+
+  @Watch('step')
+  onStepChanged (step: number | string) {
+    if (String(step) !== '3') {
+      return
+    }
+
+    this.$nextTick(() => {
+      const form = this.$refs.DynamicLocationActionInfoFormWizard as (Vue & { autoSelectContacts?: () => void }) | undefined
+      if (form?.autoSelectContacts) {
+        form.autoSelectContacts()
+      }
+    })
   }
 
   validateFormDynamicLocationActionDateForm () {